+++ praw/models/listing/listing.py
@@ -62,7 +62,7 @@ class ModNoteListing(Listing):
 
     @property
     def after(self) -> Optional[Any]:
-        """Return the next attribute or None."""
+        """Return the next attribute or ``None``."""
         if not getattr(self, "has_next_page", True):
             return None
         return getattr(self, "end_cursor", None)